requirements.txt
aicrowd-cli==0.1.15
hydra-core==1.3.2
pandas==2.0.3
torch==2.0.1
torchvision==0.15.2
albumentations==1.3.1
pillow==10.0.0
matplotlib==3.7.2
wandb==0.13.11